CPM-1151: Move CreateAttributeProgressIndicator to be used for table attribute
