+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/attribute/form/CreateAttributeButtonApp.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import SelectAttributeType from './SelectAttributeType';
 import {Button, useBooleanState} from 'akeneo-design-system';
 import {Modal} from 'akeneo-design-system';
-import {CreateAttributeProgressIndicator} from './CreateAttributeProgressIndicator';
+import {CreateAttributeProgressIndicator} from '@akeneo-pim-community/settings-ui';
 
 export type CreateAttributeButtonStepProps = {
   onClose: () => void;
@@ -38,7 +38,7 @@ const CreateAttributeButtonApp: React.FC<CreateAttributeButtonAppProps> = ({
   const [isOpen, open, close] = useBooleanState(isModalOpen);
   const [attributeData, setAttributeData] = React.useState<AttributeData>(initialData);
   const [currentStepIndex, setCurrentStepIndex] = React.useState<number>(-1);
-
+  console.log('CreateAttributeButtonApp');
   let stepsForAttributeType = steps.default;
   if (attributeData.attribute_type) {
     if (attributeData.attribute_type in steps) {
